Move command suggestion helpers to CMUXCLI+CommandSuggestions.swift
workflow-guard-tests failed because the new helpers pushed CLI/cmux.swift
47 lines over its length budget. Move unknownCommandError,
suggestedCommandName, editDistance, and topLevelCommandNames into a new
extension file (wired into the pbxproj) and ratchet the budget entry down
to the new 34499-line count. No behavior change; verified on the rebuilt
clicpy tag.
